+++ src/GraphDataInputs/expand.jl
@@ -140,14 +140,14 @@ s(n) = n == 1 ? "" : "s"
 # TODO: abstract over map/adjacency with in the same fashion as for checking?
 
 # Assuming all indices have been given.
-function to_dense_vector(map::Map{Int64,T}) where {T}
+function to_dense_vector(map::AbstractMap{Int64,T}) where {T}
     res = Vector{T}(undef, length(map))
     for (i, value) in map
         res[i] = value
     end
     res
 end
-function to_dense_vector(map::Map{Symbol,T}, index) where {T}
+function to_dense_vector(map::AbstractMap{Symbol,T}, index) where {T}
     res = Vector{T}(undef, length(map))
     for (key, value) in map
         i = index[key]
@@ -158,14 +158,14 @@ end
 export to_dense_vector
 
 # Assuming all indices are valid.
-function to_sparse_vector(map::Map{Int64,T}, n::Int64) where {T}
+function to_sparse_vector(map::AbstractMap{Int64,T}, n::Int64) where {T}
     res = spzeros(T, n)
     for (i, value) in map
         res[i] = value
     end
     res
 end
-function to_sparse_vector(map::Map{Symbol,T}, index) where {T}
+function to_sparse_vector(map::AbstractMap{Symbol,T}, index) where {T}
     res = spzeros(T, length(index))
     for (key, value) in map
         i = index[key]
@@ -173,14 +173,14 @@ function to_sparse_vector(map::Map{Symbol,T}, index) where {T}
     end
     res
 end
-function to_sparse_vector(map::BinMap{Int64}, n::Int64)
+function to_sparse_vector(map::AbstractBinMap{Int64}, n::Int64)
     res = spzeros(Bool, n)
     for i in map
         res[i] = true
     end
     res
 end
-function to_sparse_vector(map::BinMap{Symbol}, index)
+function to_sparse_vector(map::AbstractBinMap{Symbol}, index)
     res = spzeros(Bool, length(index))
     for key in map
         i = index[key]
@@ -191,20 +191,21 @@ end
 export to_sparse_vector
 
 # Accommodate slight signature variations in case an index is always used.
-function to_dense_vector(map::Map{Int64,T}, index) where {T}
+function to_dense_vector(map::AbstractMap{Int64,T}, index) where {T}
     m, i = length.((map, index))
     m == i || argerr("Cannot produce a dense vector with $m values and $i references.")
     to_dense_vector(map)
 end
-to_sparse_vector(map::Map{Int64,T}, index) where {T} = to_sparse_vector(map, length(index))
-to_sparse_vector(map::BinMap{Int64}, index) = to_sparse_vector(map, length(index))
+to_sparse_vector(map::AbstractMap{Int64,T}, index) where {T} =
+    to_sparse_vector(map, length(index))
+to_sparse_vector(map::AbstractBinMap{Int64}, index) = to_sparse_vector(map, length(index))
 
 #-------------------------------------------------------------------------------------------
 # Assuming the input is a correctly checked adjacency list,
 # expand to a sparse matrix.
 # This may require a label-to-indices mapping referred to as "index" (yup, confusing).
 
-function to_sparse_matrix(adj::Adjacency{Int64,T}, n::Int64, m::Int64) where {T}
+function to_sparse_matrix(adj::AbstractAdjacency{Int64,T}, n::Int64, m::Int64) where {T}
     res = spzeros(T, (n, m))
     for (i, list) in adj
         for (j, value) in list
@@ -213,7 +214,7 @@ function to_sparse_matrix(adj::Adjacency{Int64,T}, n::Int64, m::Int64) where {T}
     end
     res
 end
-function to_sparse_matrix(adj::Adjacency{Symbol,T}, i_index, j_index) where {T}
+function to_sparse_matrix(adj::AbstractAdjacency{Symbol,T}, i_index, j_index) where {T}
     res = spzeros(T, (length(i_index), length(j_index)))
     for (ikey, list) in adj
         i = i_index[ikey]
@@ -224,7 +225,7 @@ function to_sparse_matrix(adj::Adjacency{Symbol,T}, i_index, j_index) where {T}
     end
     res
 end
-function to_sparse_matrix(adj::BinAdjacency{Int64}, n::Int64, m::Int64)
+function to_sparse_matrix(adj::AbstractBinAdjacency{Int64}, n::Int64, m::Int64)
     res = spzeros(Bool, (n, m))
     for (i, list) in adj
         for j in list
@@ -233,7 +234,7 @@ function to_sparse_matrix(adj::BinAdjacency{Int64}, n::Int64, m::Int64)
     end
     res
 end
-function to_sparse_matrix(adj::BinAdjacency{Symbol}, i_index, j_index)
+function to_sparse_matrix(adj::AbstractBinAdjacency{Symbol}, i_index, j_index)
     res = spzeros(Bool, (length(i_index), length(j_index)))
     for (ikey, list) in adj
         i = i_index[ikey]
@@ -247,9 +248,9 @@ end
 export to_sparse_matrix
 
 # Accommodate slight signature variations in case an index is always used.
-to_sparse_matrix(map::Adjacency{Int64,T}, i_index, j_index) where {T} =
+to_sparse_matrix(map::AbstractAdjacency{Int64,T}, i_index, j_index) where {T} =
     to_sparse_matrix(map, length(i_index), length(j_index))
-to_sparse_matrix(map::BinAdjacency{Int64}, i_index, j_index) =
+to_sparse_matrix(map::AbstractBinAdjacency{Int64}, i_index, j_index) =
     to_sparse_matrix(map, length(i_index), length(j_index))
 
 # ==========================================================================================
@@ -321,7 +322,7 @@ export @expand_from_col_if_vector
 macro to_dense_vector_if_map(var::Symbol, index)
     var, index = esc.((var, index))
     quote
-        $var isa Union{OrderedDict,OrderedSet} && ($var = to_dense_vector($var, $index))
+        $var isa Union{AbstractDict,AbstractSet} && ($var = to_dense_vector($var, $index))
     end
 end
 export @to_dense_vector_if_map
@@ -329,7 +330,7 @@ export @to_dense_vector_if_map
 macro to_sparse_vector_if_map(var::Symbol, index)
     var, index = esc.((var, index))
     quote
-        $var isa Union{OrderedDict,OrderedSet} && ($var = to_sparse_vector($var, $index))
+        $var isa Union{AbstractDict,AbstractSet} && ($var = to_sparse_vector($var, $index))
     end
 end
 export @to_sparse_vector_if_map
@@ -337,7 +338,8 @@ export @to_sparse_vector_if_map
 macro to_sparse_matrix_if_adjacency(var::Symbol, i_index, j_index)
     var, i_index, j_index = esc.((var, i_index, j_index))
     quote
-        $var isa Union{OrderedDict{<:Any,<:OrderedDict},OrderedDict{<:Any,<:OrderedSet}} &&
+        $var isa
+        Union{AbstractDict{<:Any,<:AbstrsactDict},AbstractDict{<:Any,<:AbstractSet}} &&
             ($var = to_sparse_matrix($var, $i_index, $j_index))
     end
 end