+++ test/user/03-components.jl
@@ -25,6 +25,7 @@ only = [
     "./data_components/temperature.jl"
     "./data_components/hill_exponent.jl"
     "./data_components/growth_rate.jl"
+    "./data_components/efficiency.jl"
 ] # Only run these if specified.
 if isempty(only)
     for subfolder in ["./data_components", "./code_components"]