+++ test/user/data_components/species.jl
@@ -1,21 +1,31 @@
 @testset "Species components." begin
 
-    m = Model(Species([:a, :b, :c]))
+    base = Model()
 
+    # From names.
+    sp = Species([:a, :b, :c])
+    m = base + sp
     @test m.richness == 3
     @test m.S == 3
     @test m.species.richness == 3
     @test m.species.number == 3
     @test m.species.index == Dict(:a => 1, :b => 2, :c => 3)
     @test m.species.names == [:a, :b, :c]
+    @test typeof(sp) == Species.Names
 
     # Get a closure to convert index to label.
     lab = m.species.label
     @test lab(1) == :a
     @test lab.([1, 2, 3]) == [:a, :b, :c]
 
-    # Default names.
-    @test Model(Species(3)).species.names == [:s1, :s2, :s3]
+    # From a number of species (default names generated).
+    sp = Species(3)
+    m = base + sp
+    @test m.species.names == [:s1, :s2, :s3]
+    @test typeof(sp) == Species.Number
+
+    #---------------------------------------------------------------------------------------
+    # Guards.
 
     @sysfails(
         Model(Species([:a, :b, :a])),
@@ -27,7 +37,6 @@
         "Invalid index (3) when there are 2 species names."
     )
 
-    # Cannot query without the component.
     @sysfails(
         Model().richness,
         Property(richness, "Component $(EN._Species) is required to read this property."),